backend/requirements.txt
Django>=5.0

graphene-django>=3.2

psycopg2-binary>=2.9    
python-dotenv>=1.0      

django-cors-headers>=4.3 

pytest>=8.0
pytest-django>=4.8
mixer>=7.2
django-graphql-jwt>=0.4.0
